# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the backend API for QuokkaQ, rewritten in Go.",
        "title": "QuokkaQ Go Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3001",
    "basePath": "/",
    "paths": {
...
        "/counters": {
            "post": {
                "description": "Creates a new counter for a unit",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "counters"
                ],
                "summary": "Create a new counter",
                "parameters": [
                    {
                        "description": "Counter Data",
                        "name": "counter",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Counter"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Counter"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/counters/{id}": {
            "get": {
                "description": "Retrieves a specific counter by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "counters"
                ],
                "summary": "Get a counter by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Counter ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Counter"
                        }
                    },
                    "404": {
                        "description": "Counter not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates an existing counter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "counters"
                ],
                "summary": "Update a counter",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Counter ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Counter Data",
                        "name": "counter",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Counter"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Counter"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a counter by its ID",
                "tags": [
                    "counters"
                ],
                "summary": "Delete a counter",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Counter ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/counters/{id}/force-release": {
            "post": {
                "description": "Force releases a counter (supervisor)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "counters"
                ],
                "summary": "Force release counter",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Counter ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Counter not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/counters/{id}/occupy": {
            "post": {
                "description": "Sets a counter as occupied by the current user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "counters"
                ],
                "summary": "Occupy counter",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Counter ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Counter"
                        }
                    },
                    "404": {
                        "description": "Counter not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/counters/{id}/release": {
            "post": {
                "description": "Releases a counter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "counters"
                ],
                "summary": "Release counter",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Counter ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Counter"
                        }
                    },
                    "404": {
                        "description": "Counter not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/units/{unitId}/counters": {
            "get": {
                "description": "Retrieves all counters for a specific unit",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "counters"
                ],
                "summary": "Get counters by unit",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Unit ID",
                        "name": "unitId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Counter"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Counter": {
            "type": "object",
            "properties": {
                "assignedTo": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "unitId": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}